Move no reference test
